[feat] Add non-production mode for runner
In non-production mode we do not create multiple directory and we do not
create reference directories

+++ runner.sh
@@ -1,13 +1,20 @@
 #!/bin/bash
 
+## runner FILE PRODUCTION(yes|null)
+
 TIMESTAMP=`date '+%s'`
 OUTPUT="output-${TIMESTAMP}"
 
-REFERENCE=`ls -td -- output-* | head -n1`
+REFERENCE=`ls -td -- output-* | grep -v output-tmp | head -n1`
 
 if [ -z "$1" ]; then
     echo "You have to enter file with documents to parse"
     exit 1
+fi
+
+if [ -z "$2" ]; then
+    # we are not in production mode
+    OUTPUT="output-tmp"
 fi
 
 mkdir -p ${OUTPUT}
@@ -18,8 +25,11 @@ find *pretty -type f -exec md5sum {} + | sort -k 2 > tree.md5
 cd ..
 
 diff -c1 ${OUTPUT}/tree.md5 ${REFERENCE}/tree.md5
+wc -l ${OUTPUT}/tree.md5 $1 | head -n2
 
 if [ "$?" -eq "0" ]; then
     # results are same, so there is no need to preserve them
-    rm -rf ${OUTPUT}
+    if [ ! -z "$2"]; then
+        rm -rf ${OUTPUT}
+    fi
 fi
